Deduplicate Drop Forge nav helper
